news_and_authorization_service/requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg==3.1.19
psycopg-binary==3.1.19
PyJWT==2.8.0
python-dotenv==1.0.1
